2020-08-21-Auto Mail configuration

+++ Attendance/Forms/frmUploadShiftSchedule.cs
@@ -554,20 +554,40 @@ namespace Attendance.Forms
                                 }//end if WO
                                 else
                                 {
-                                    try
-                                    {
-                                        sql = "Delete from MastLeaveSchedule Where EmpUnqID = '" + Emp.EmpUnqID + "' " +
-                                                    " and tDate = '" + date.ToString("yyyy-MM-dd") + "' and SchLeave ='WO'";
 
-                                        cmd = new SqlCommand(sql,cn,trn);
-                                        cmd.ExecuteNonQuery();
-                                    }
-                                    catch (Exception ex)
+                                    LeavPos = false;
+                                    foreach (LeaveData t in leave)
                                     {
-                                        dr["Remarks"] = dr["Remarks"].ToString() + Environment.NewLine + ex.ToString();
-                                        brkflg = true;
-                                        break;
+                                        //check if date is fall between leave posted
+                                        if (date >= t.FromDt && date <= t.ToDt)
+                                        {
+                                            LeavPos = true;
+                                            break;
+                                        }
                                     }
+
+                                    if (!LeavPos)
+                                    {
+                                        try
+                                        {
+                                            sql = "Delete from MastLeaveSchedule Where EmpUnqID = '" + Emp.EmpUnqID + "' " +
+                                                        " and tDate = '" + date.ToString("yyyy-MM-dd") + "' and SchLeave ='WO'";
+
+                                            cmd = new SqlCommand(sql, cn, trn);
+                                            cmd.ExecuteNonQuery();
+                                        }
+                                        catch (Exception ex)
+                                        {
+                                            dr["Remarks"] = dr["Remarks"].ToString() + Environment.NewLine + ex.ToString();
+                                            brkflg = true;
+                                            break;
+                                        }
+
+                                    }
+                                    
+                                    
+                                    
+                                    
                                 }
 
                             } //end of eachday
